Saving changes before reloading project (IDE issues)
